Fixing integer math

+++ C0_bug1_numpy_simplicialComplex_encoder_for_array_of_reals_as_multiset.py
@@ -245,7 +245,7 @@ def map_Delta_k_to_the_n_to_c_dc_pairs(#n=3,k=3,  # Only need n and/or k if doin
     return c_dc_pairs
 
 def pr(r, big_n):
-    return np.power(r, np.arange(1, big_n+1)) # TODO: Ask PK-H whether there is a better final step than pr
+    return np.power(np.float64(r), np.arange(1, big_n+1)) # TODO: Ask PK-H whether there is a better final step than pr
 
 def map_Delta_k_to_the_n_to_c_l_dc_triples(n, k, delta):
     c_dc_pairs = map_Delta_k_to_the_n_to_c_dc_pairs(delta=delta)
